Move vpn tests to JVM (#2250)
