+++ app-tracking-protection/vpn-impl/build.gradle
@@ -30,6 +30,12 @@ android {
     anvil {
         generateDaggerFactories = true // default is false
     }
+
+    testOptions {
+        unitTests {
+            includeAndroidResources = true
+        }
+    }
 }
 
 ext {
@@ -82,11 +88,6 @@ dependencies {
     implementation Google.dagger
 
     implementation KotlinX.coroutines.core
-    androidTestImplementation (KotlinX.coroutines.test) {
-        // https://github.com/Kotlin/kotlinx.coroutines/issues/2023
-        // conflicts with mockito due to direct inclusion of byte buddy
-        exclude group: "org.jetbrains.kotlinx", module: "kotlinx-coroutines-debug"
-    }
 
     testImplementation (KotlinX.coroutines.test) {
         // https://github.com/Kotlin/kotlinx.coroutines/issues/2023
@@ -105,14 +106,11 @@ dependencies {
     implementation AndroidX.room.runtime
     implementation AndroidX.room.rxJava2
     implementation AndroidX.room.ktx
-    testImplementation AndroidX.room.testing
-    androidTestImplementation AndroidX.room.testing
 
     // WorkManager
     implementation AndroidX.work.runtimeKtx
     implementation "androidx.work:work-multiprocess:$workManager"
     testImplementation AndroidX.work.testing
-    androidTestImplementation AndroidX.work.testing
     implementation AndroidX.work.rxJava2
 
     implementation Square.retrofit2.converter.moshi
@@ -127,24 +125,5 @@ dependencies {
     testImplementation AndroidX.archCore.testing
     testImplementation 'app.cash.turbine:turbine:_'
     testImplementation Testing.robolectric
-
-    androidTestImplementation AndroidX.archCore.testing
-    androidTestImplementation 'app.cash.turbine:turbine:_'
-
-    androidTestImplementation AndroidX.test.runner
-    androidTestImplementation AndroidX.test.rules
-
-    androidTestImplementation "org.mockito:mockito-android:_"
-    androidTestImplementation "org.mockito.kotlin:mockito-kotlin:_"
-
     testImplementation project(path: ':common-test')
-    androidTestImplementation project(path: ':common-test')
-}
-
-fulladleModuleConfig {
-    maxTestShards.set(1)
-}
-
-tasks.register('androidTestsBuild') {
-    dependsOn 'assembleDebugAndroidTest'
 }
